betty

+++ search_algorithms/0-linear.c
@@ -9,7 +9,7 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	long unsigned int i;
+	unsigned int i;
 
 	if (array == NULL)
 	{
@@ -18,7 +18,7 @@ int linear_search(int *array, size_t size, int value)
 
 	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
 			return (i);
